Реорганизовал код orders_service: вынес обработку маршрутов в routes.py, добавил функции работы с Redis в redis_service.py. Упрощён main.py, теперь использует router для обработки запросов.
