+++ orders_service/app/redis_service.py
@@ -1,0 +1,8 @@
+import aioredis
+
+async def get_redis():
+    redis = aioredis.from_url("redis://redis:6379", encoding="utf-8", decode_responses=True)
+    try:
+        yield redis
+    finally:
+        await redis.close()
